==> notebooks/0_etl_json_to_parquet.ipynb <==
import polars as pl
import pickle
from hgm.config import DATA_DIR

DATA_FILEPATH = r'C:\Users\jrnas\Projects\Personal\hgm\league_files\analytics\ZGMH_Analytics__30_man__2099_draft_lottery(1).json'

json_data = pl.read_json(DATA_FILEPATH)
json_data.select('players').explode('players').unnest('players').write_parquet(DATA_DIR / 'raw' / 'players.parquet')
json_data.select('teams').explode('teams').unnest('teams').write_parquet(DATA_DIR / 'raw' / 'teams.parquet')
with open(DATA_DIR / 'raw' / 'game_settings.pkl', 'wb') as file:
    pickle.dump(json_data.select('gameAttributes').item(), file)

json_data.head()
