==> notebooks/1_calculate_sals.ipynb <==
import polars as pl
import pickle

from hgm.config import MODELS_DIR, DATA_DIR

players_raw = pl.scan_parquet(DATA_DIR / 'raw' / 'players.parquet')

with open(DATA_DIR / 'raw' / 'game_settings.pkl', 'rb') as file:
    game_settings = pickle.load(file)

player_ratings = (
    players_raw
    .with_columns(
        pl.concat_str(pl.col('firstName'), pl.lit(' '), pl.col('lastName')).alias('player'),
    )
    .explode('ratings')
    .unnest('ratings')
    .with_columns(
        age=pl.col('season') - pl.col('born').struct.field('year')
    )
    .unique(['pid', 'season'])
    .sort(['pid', 'season'])
    .select('player', 'pid', 'tid', 'season', 'age', 'pos', 'ovr')
    .collect()
)

player_salaries = (
    players_raw
    .select('pid', 'salaries')
    .explode('salaries')
    .unnest('salaries')
    .group_by(['pid', 'season'])
    .agg(salary=pl.col('amount').last())
    .sort('pid', 'season')
    .collect()
)

new_contracts = (
    player_salaries
    .with_columns(
        contract_year=pl.col('salary') != pl.col('salary').shift(-1).over('pid'),
    )
    .filter(pl.col('contract_year'))
    .select(
        pl.col('pid'),
        pl.col('season'),
        pl.col('salary').alias('current_salary') / 1000,
        pl.col('salary').shift(-1).over('pid').alias('next_salary') / 1000,
    )
    .drop_nulls()
)

contract_data = (
    new_contracts
    .join(player_ratings, on=['pid', 'season'], how='left')
    .select(
        'pid', 'age', 'ovr', 'current_salary', 'next_salary',
        pl.col('pos').replace({
            'C': 1,
            'W': 2,
            'D': 3,
            'G': 4
        }).cast(pl.Int32).alias('pos_map')
    )
)

from sklearn.model_selection import train_test_split
from xgboost import XGBRegressor
import pickle

features = ['pos_map', 'age', 'ovr', 'current_salary']
target = 'next_salary'

# Split the data into training and validation sets
X_train, X_val, y_train, y_val = (
    train_test_split(
        contract_data.select(features),
        contract_data.select(target),
        test_size=0.2,
        random_state=42
    )
)

model = XGBRegressor(
    early_stopping_rounds=5,
    learning_rate=0.1,
    random_state=42,
)

# Fit the model using the training data and validate using the validation data
model.fit(X_train, y_train, eval_set=[(X_val, y_val)])

with open(MODELS_DIR / 'salary' / 'xgboost_salary.pkl', 'wb') as file:
    pickle.dump(model, file)
